functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core

azure-ai-textanalytics~=5.3.0
azure-data-tables~=12.5.0
azure-storage-queue~=12.9.0
beautifulsoup4~=4.11.1
feedparser~=6.0.8
requests~=2.27.1
